cassandra_example/pyproject.toml
[tool.poetry]
name = "cassandra-example"
version = "0.1.0"
description = ""
authors = ["Lincoln Nguyen <61459043+lincolnnguyen18@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "cassandra_example"}]

[tool.poetry.dependencies]
python = "^3.11"
cassandra-driver = "^3.26.0"
strawberry-graphql = {extras = ["debug-server"], version = "^0.169.0"}


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"
pytest = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
